Clean up zones on teleport
